refactor(server): moved internal/utils to pkg
also ran go mod tidy
